refactor and delete annotation schema-related code
I may be removing important functionality, but I don't see the use case
for providing a user-defined annotation schema, since it's closely tied
to the trained model. Remove the code and parameters for now. Add TODOs
related to proper handling of the annotation schema when fine-tuning.

Use logic from the connlup package instead of the custom code to read
connlu files for the annotation schema.
