+++ TODO.md
@@ -3,6 +3,8 @@
 Next: How does Trankit save memory by only loading XLM-Roberta once?
 Then: build Flask server - prediction is currently hardcoded to read from and write to conllu files - need to load model in an init method and use it for subsequent requests
 
+-   Remove 26.1, etc. from Conllu data. We can't predict these right now.
+
 Update readme with new build info: poetry, poe commands, pytest, verification, pre-commit, etc.
 
 -   Try XLM-RoBERTa-XL, which is 2 years newer than xlm-roberta
